Code review rework
