+++ src/test/java/org/pac4j/jax/rs/RestEasyUndertowServletTest.java
@@ -1,6 +1,5 @@
 package org.pac4j.jax.rs;
 
-import org.junit.Ignore;
 import org.pac4j.jax.rs.rules.RestEasyUndertowServletRule;
 import org.pac4j.jax.rs.rules.SessionContainerRule;
 
@@ -17,52 +16,4 @@ public class RestEasyUndertowServletTest extends AbstractSessionTest {
         return new RestEasyUndertowServletRule();
     }
 
-    // TODO we don't have injection yet for something else than Jersey!
-    @Ignore
-    @Override
-    public void testInject() {
-        // do nothing
-    }
-
-    // TODO we don't have injection yet for something else than Jersey!
-    @Ignore
-    @Override
-    public void directInject() {
-        // do nothing
-    }
-
-    // TODO we don't have injection yet for something else than Jersey!
-    @Ignore
-    @Override
-    public void directInjectNoAuth() {
-        // do nothing
-    }
-
-    // TODO we don't have injection yet for something else than Jersey!
-    @Ignore
-    @Override
-    public void directInjectSkipFail() {
-        // do nothing
-    }
-
-    // TODO we don't have injection yet for something else than Jersey!
-    @Ignore
-    @Override
-    public void directInjectSkipOk() {
-        // do nothing
-    }
-
-    // TODO we don't have injection yet for something else than Jersey!
-    @Ignore
-    @Override
-    public void directInjectManagerAuth() {
-        // do nothing
-    }
-
-    // TODO we don't have injection yet for something else than Jersey!
-    @Ignore
-    @Override
-    public void directInjectManagerNoAuth() {
-        // do nothing
-    }
 }
