+++ src/test/java/org/pac4j/jax/rs/rules/RestEasyUndertowServletRule.java
@@ -15,6 +15,7 @@ import org.jboss.resteasy.test.TestPortProvider;
 import org.junit.rules.ExternalResource;
 import org.pac4j.jax.rs.features.JaxRsConfigProvider;
 import org.pac4j.jax.rs.features.Pac4JSecurityFeature;
+import org.pac4j.jax.rs.resteasy.features.Pac4JProfileInjectorFactory;
 import org.pac4j.jax.rs.servlet.features.ServletJaxRsContextFactoryProvider;
 
 import io.undertow.server.session.SessionCookieConfig;
@@ -34,7 +35,10 @@ public class RestEasyUndertowServletRule extends ExternalResource implements Ses
 
         @Override
         public Set<Object> getSingletons() {
-            return Sets.newLinkedHashSet(new JaxRsConfigProvider(getConfig()), new ServletJaxRsContextFactoryProvider(),
+            return Sets.newLinkedHashSet(
+                    new JaxRsConfigProvider(getConfig()),
+                    new ServletJaxRsContextFactoryProvider(),
+                    new Pac4JProfileInjectorFactory(),
                     new Pac4JSecurityFeature());
         }
     }
